feat:Client work in progress
