+++ client.h
@@ -1,0 +1,27 @@
+
+
+#ifndef CLIENT_H
+#define CLIENT_H
+
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <unistd.h>
+#include <curses.h>
+// Sockets libraries
+#include <netdb.h>
+#include <arpa/inet.h>
+// poll
+#include <sys/poll.h>
+// Custom libraries
+#include "protocol.h"
+#include "sockets.h"
+#include "fatal_error.h"
+
+
+void purgeStdin();
+
+char* getUsername();
+
+
+#endif
